Add to contributing guide: standards and finding something to work
Recently I wanted to give feedback on a commit message and I realized we
didn't have anything in our CONTRIBUTING.md about it!

Also added the section on finding issues b/c we'd really like to
encourage folks to get involved with Kaniko.

(This is based on `knative/build-pipeline`'s CONTRIBUTING.md:
https://github.com/knative/build-pipeline/blob/master/CONTRIBUTING.md)