Don't recreate cached layers
This caching optimization reuses existing layer from a cache image for a
RUN command instead of unpacking it and recreating.  It also helps with
multi-stage build caching issues - when recreating cached layers, some
attributes like timestamps are not recreated and the digest for the new
first stage image was different, breaking caching for subsequent stages.