Consistently handle the case where config.WorkingDir==""
So given that we os.Chdir("/") at startup, I can't figure out how not
handling this can break RUN or ADD, but adjusting them to explicitly
handle this (in the same way as COPY already does) is easier than proving
that they don't need to.

WORKDIR, on the other hand, definitely needs to handle this explicitly;
many images don't have Config.WorkingDir set; such as ubuntu:18.04.

For example, given the Dockerfile:

    FROM ubuntu:18.04
    WORKDIR ./app

Kaniko produces an image that cannot be run normally:

    $ docker run --rm -it IMAGENAME
    ...
    Status: Downloaded newer image for IMAGENAME
    docker: Error response from daemon: OCI runtime create failed: Cwd must be an absolute path: unknown.