Update readme insecure flags (#1811)
* Clarify usage of --skip-tls-verify-registry argument

Referring to --skip-tls-verify-registry as a "flag" is misleading since
command line flags usually require no value. Thus, the documentation
is updated to match the behavior of the command line argument.

* Clarify usage of --insecure-registry argument

Referring to --insecure-registry as a "flag" is misleading since
command line flags usually require no value. Thus, the documentation
is updated to match the behavior of the command line argument.