Fix bug with log disabling
Prior to this change, Kaniko disabled logging for its ecr-login
dependency by redirecting the logger's output to ioutil.Discard.
However, since that used the logrus standard logger, this had the effect
of completely disabling (redirecting to Discard) _all_ logging done by
Kaniko.

This caused integratiion test failures, and left any user of Kaniko
without any logging.

This fix updates the dependency on ecr-login to include a fix for this
issue, and updates to the new API introduced in that fix.