Download docker-credential-gcr from release artifacts
As described in #2539, using `go install` with a pinned hash version it installs `1.5.1` in all cases.
This seems to be an issue on the upstream project https://github.com/GoogleCloudPlatform/docker-credential-gcr and how go proxy deals with versions.

As a temporary solution, we download the compiled version from the release artifacts.

Closes #2539

Signed-off-by: Ferran Vidal <ferran.vidal.p@gmail.com>