Speedup --reproducible layer stripping
Fixes #862, may mitigate #1960

The layerTime function returns a layer backed by an in-memory buffer, which means during
the stripping of timestamps the entire image is loaded into memory. Additionally, it runs
gzip after the layer is created, resulting in even an even larger in-memory blob.

This commit changes this method to use a temporary file instead of an in-memory buffer,
and to use gzip compression while writing to this layer file, instead of compressing during read.

Signed-off-by: bh <bert@transtrend.com>